Penalities

+++ database/migrations/2022_07_29_135623_create_lump_sum_returns_table.php
@@ -20,7 +20,7 @@ class CreateLumpSumReturnsTable extends Migration
             $table->unsignedBigInteger('financial_year_id');
             $table->integer('quarter');
 
-            $table->decimal('amount')->default(0);
+            $table->integer('amount')->default(0);
             $table->integer('receipt_no');
             $table->integer('edited_count')->default(0);
             $table->string('currency');
